requirements.txt
# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-dotenv==1.0.1

# ── Web scraping ──────────────────────────────────────────────────────────────
selenium==4.18.1
beautifulsoup4==4.12.3
requests==2.31.0
webdriver-manager==4.0.1       # automatic ChromeDriver management (optional)

# ── Financial data ────────────────────────────────────────────────────────────
yfinance==0.2.37

# ── Data processing (Sprint 2) ────────────────────────────────────────────────
pandas==2.2.1
numpy==1.26.4

# ── Machine learning (Sprint 3) ───────────────────────────────────────────────
scikit-learn==1.4.1
xgboost==2.0.3
joblib==1.3.2

# ── Visualization & notebooks (Sprint 2) ──────────────────────────────────────
matplotlib==3.8.3
seaborn==0.13.2
jupyter

# ── NLP & Sentiment (Sprint 4) ────────────────────────────────────────────────
transformers==4.39.3
torch==2.2.2

# ── Progress ──────────────────────────────────────────────────────────────────
tqdm==4.66.2
